+++ packages/cloud-cognitive/src/components/APIKeyModal/APIKeyModal.test.js
@@ -162,7 +162,6 @@ describe(name, () => {
       hasDownloadLink: false,
       apiKey: '',
     };
-
     const { rerender, getByPlaceholderText, getByText } = render(
       <APIKeyModal {...props} />
     );