+++ packages/cloud-cognitive/src/components/AboutModal/AboutModal.test.js
@@ -5,6 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+// cspell:words grafana
+
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
@@ -88,6 +90,21 @@ const renderComponent = ({ ...rest }) =>
   );
 
 describe(componentName, () => {
+  const { ResizeObserver } = window;
+
+  beforeEach(() => {
+    window.ResizeObserver = jest.fn().mockImplementation(() => ({
+      observe: jest.fn(),
+      unobserve: jest.fn(),
+      disconnect: jest.fn(),
+    }));
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+    window.ResizeObserver = ResizeObserver;
+  });
+
   it('renders a component AboutModal', () => {
     renderComponent();
     expect(screen.getByRole('presentation')).toHaveClass(blockClass);
@@ -181,7 +198,7 @@ describe(componentName, () => {
   });
 
   it('adds additional properties to the containing node', () => {
-    renderComponent({ 'data-testid': dataTestId });
+    renderComponent({ 'data-test-id': dataTestId });
     screen.getByTestId(dataTestId);
   });