+++ packages/cloud-cognitive/src/components/CreateTearsheet/CreateTearsheet.js
@@ -24,7 +24,7 @@ import {
   SideNavLink,
 } from 'carbon-components-react/lib/components/UIShell';
 import cx from 'classnames';
-import ReactResizeDetector from 'react-resize-detector';
+import { useResizeDetector } from 'react-resize-detector';
 import wrapFocus from '../../global/js/utils/wrapFocus';
 import { TearsheetShell } from '../Tearsheet/TearsheetShell';
 import { carbon, pkg } from '../../settings';
@@ -72,6 +72,7 @@ export let CreateTearsheet = forwardRef(
     const [isSubmitting, setIsSubmitting] = useState(false);
     const [activeSectionIndex, setActiveSectionIndex] = useState(0);
     const previousState = usePreviousValue({ currentStep, open });
+    const contentRef = useRef();
 
     // set current step to 1 upon tearsheet opening, in order
     // to get the auto focus on the first step.
@@ -101,35 +102,43 @@ export let CreateTearsheet = forwardRef(
           isTearsheetStep(child)
         );
         let tearsheetSectionComponents = [];
-        tearsheetStepComponents.forEach((child) => {
-          // we have received an array of children, lets check to see that each child is
-          // a CreateTearsheetSection component before adding it to tearsheetSectionComponents
-          if (shouldViewAll && child.props.children.length) {
-            // only a string was provided as children of CreateTearsheetStep, this is not permitted when using view all toggle
+        tearsheetStepComponents.forEach((child, index) => {
+          // We have received children for a TearsheetStep
+          if (shouldViewAll && typeof child.props.children !== 'undefined') {
+            // Only a string was provided as children of CreateTearsheetStep, this is not permitted when using view all toggle
             if (typeof child.props.children === 'string') {
               console.warn(
                 `${componentName}: You must have at least one CreateTearsheetSection component in a CreateTearsheetStep when using the 'includeViewAllToggle' prop.`
               );
             } else {
-              child.props.children.forEach((stepChild) => {
-                if (isTearsheetSection(stepChild)) {
-                  tearsheetSectionComponents.push(stepChild);
+              // The TearsheetStep has an array of children, lets check each one to see if it is a TearsheetSection
+              if (child.props.children.length) {
+                child.props.children.forEach((stepChild) => {
+                  if (isTearsheetSection(stepChild)) {
+                    tearsheetSectionComponents.push(stepChild);
+                  }
+                });
+              } else {
+                // The TearsheetStep only has a single React element as a child, lets check to see if it is a TearsheetSection
+                if (isTearsheetSection(child.props.children)) {
+                  tearsheetSectionComponents.push(child.props.children);
                 }
-              });
-              // if there are fewer CreateTearsheetSection components than CreateTearsheetStep components
-              // it means that each CreateTearsheetStep does not have at least one CreateTearsheetSection
-              // this is not permitted when using view all toggle
-              if (
-                tearsheetSectionComponents.length <
-                tearsheetStepComponents.length
-              ) {
-                console.warn(
-                  `${componentName}: You must have at least one CreateTearsheetSection component in a CreateTearsheetStep when using the 'includeViewAllToggle' prop.`
-                );
               }
             }
           }
-          // we have received a single child element, lets check to see that it is
+          // If there are fewer CreateTearsheetSection components than CreateTearsheetStep components
+          // it means that each CreateTearsheetStep does not have at least one CreateTearsheetSection
+          // this is not permitted when using view all toggle
+          if (
+            tearsheetSectionComponents.length <
+              tearsheetStepComponents.length &&
+            index === tearsheetStepComponents.length - 1 // wait until we've finished checking each TearsheetStep before giving a warning
+          ) {
+            console.warn(
+              `${componentName}: You must have at least one CreateTearsheetSection component in a CreateTearsheetStep when using the 'includeViewAllToggle' prop.`
+            );
+          }
+          // We have received a single child element, lets check to see that it is
           // a CreateTearsheetSection component, if it is not we should add a console
           // warning, as each CreateTearsheetStep required at least one CreateTearsheetSection,
           // when using the view all toggle
@@ -169,14 +178,21 @@ export let CreateTearsheet = forwardRef(
       const isSubmitDisabled = () => {
         let step = 0;
         let submitDisabled = false;
+        let viewAllSubmitDisabled = false;
         const tearsheetSteps = getTearsheetSteps();
         tearsheetSteps.forEach((child) => {
           step++;
           if (currentStep === step) {
             submitDisabled = child.props.disableSubmit;
           }
+          if (shouldViewAll && child.props.disableSubmit) {
+            viewAllSubmitDisabled = true;
+          }
         });
-        return submitDisabled;
+        if (!shouldViewAll) {
+          return submitDisabled;
+        }
+        return viewAllSubmitDisabled;
       };
       const handleNext = async () => {
         setIsSubmitting(true);
@@ -215,6 +231,7 @@ export let CreateTearsheet = forwardRef(
         const buttons = [];
         if (total > 1 && !shouldViewAll) {
           buttons.push({
+            key: 'create-tearsheet-action-button-back',
             label: backButtonText,
             onClick: () => setCurrentStep((prev) => prev - 1),
             kind: 'secondary',
@@ -222,11 +239,13 @@ export let CreateTearsheet = forwardRef(
           });
         }
         buttons.push({
+          key: 'create-tearsheet-action-button-cancel',
           label: cancelButtonText,
           onClick: onUnmount,
           kind: shouldViewAll ? 'secondary' : 'ghost',
         });
         buttons.push({
+          key: 'create-tearsheet-action-button-submit',
           label: shouldViewAll
             ? submitButtonText
             : currentStep < total
@@ -389,12 +408,15 @@ export let CreateTearsheet = forwardRef(
       );
     };
 
-    // renders all children (CreateTearsheetSteps and regular children elements)
+    // renders all children (CreateTearsheetSteps and regular child elements)
     const renderChildren = (childrenElements) => {
       let step = 0;
       const childrenArray = Array.isArray(childrenElements)
         ? childrenElements
         : [childrenElements];
+      const indexOfLastTearsheetStep = childrenArray
+        .map((el) => el?.props?.type)
+        .lastIndexOf(CREATE_TEARSHEET_STEP);
       return (
         <>
           {' '}
@@ -424,7 +446,10 @@ export let CreateTearsheet = forwardRef(
                     {renderStepTitle(stepIndex)}
                   </h4>
                 )}
-                {renderStepChildren(child.props.children)}
+                {renderStepChildren(
+                  child.props.children,
+                  indexOfLastTearsheetStep === step - 1
+                )}
               </>
             );
           })}
@@ -432,16 +457,24 @@ export let CreateTearsheet = forwardRef(
       );
     };
 
-    const renderStepChildren = (stepChildren) => {
-      const childrenArray = Array.isArray(stepChildren)
-        ? stepChildren
-        : [stepChildren];
+    const renderStepChildren = (
+      tearsheetStepComponent,
+      isLastTearsheetStep
+    ) => {
+      const tearsheetStepComponents = Array.isArray(tearsheetStepComponent)
+        ? tearsheetStepComponent
+        : [tearsheetStepComponent];
       return (
         <>
-          {childrenArray.map((child, index) => {
+          {tearsheetStepComponents.map((child, index) => {
             if (!isTearsheetSection(child)) {
               return child;
             }
+            // Needed to be able to not render the divider
+            // line on the last section of the last step
+            const isLastSectionOfLastStep =
+              isLastTearsheetStep &&
+              tearsheetStepComponents.length - 1 === index;
             return React.cloneElement(
               child,
               {
@@ -461,7 +494,7 @@ export let CreateTearsheet = forwardRef(
                   </h4>
                 )}
                 {child}
-                {shouldViewAll && (
+                {shouldViewAll && !isLastSectionOfLastStep && (
                   <span className={`${blockClass}__section--divider`} />
                 )}
               </>
@@ -563,7 +596,7 @@ export let CreateTearsheet = forwardRef(
       );
     };
 
-    const handleResize = useCallback(() => {
+    const handleResize = () => {
       const createTearsheetOuterElement = document.querySelector(
         `.${blockClass} .${carbon.prefix}--modal-container`
       );
@@ -576,7 +609,7 @@ export let CreateTearsheet = forwardRef(
         `--${blockClass}--total-width`,
         `${totalTearsheetWidth}px`
       );
-    }, []);
+    };
 
     // track scrolling/intersection of create sections so that we know
     // which section is active (updates the SideNavItems `isActive` prop)
@@ -616,44 +649,41 @@ export let CreateTearsheet = forwardRef(
       }
     }, [shouldViewAll]);
 
+    useResizeDetector({
+      handleWidth: true,
+      onResize: handleResize,
+      targetRef: contentRef,
+    });
+
     return (
-      <ReactResizeDetector handleWidth={true} onResize={handleResize}>
-        {/*
-          ReactResizeDetector needs the TearsheetShell to be wrapped inside a DOM
-          element to avoid `targetRef` being applied directly as an attribute to
-          TearsheetShell.
-         */}
-        <div>
-          <TearsheetShell
-            {...rest}
-            actions={createTearsheetActions}
-            className={cx(blockClass, className)}
-            closeIconDescription={'Close icon'}
-            description={description}
-            hasCloseIcon={false}
-            influencer={
-              <>
-                {renderProgressSteps(children)}
-                {includeViewAllToggle && renderViewAllToggle()}
-              </>
-            }
-            influencerPosition="left"
-            influencerWidth="narrow"
-            label={label}
-            onClose={onClose}
-            open={open}
-            size="wide"
-            title={title}
-            verticalPosition={verticalPosition}
-            ref={ref}>
-            <div
-              className={`${blockClass}__multi-step-panel-content`}
-              onBlur={handleBlur}>
-              {renderChildren(children)}
-            </div>
-          </TearsheetShell>
+      <TearsheetShell
+        {...rest}
+        actions={createTearsheetActions}
+        className={cx(blockClass, className)}
+        description={description}
+        hasCloseIcon={false}
+        influencer={
+          <>
+            {renderProgressSteps(children)}
+            {includeViewAllToggle && renderViewAllToggle()}
+          </>
+        }
+        influencerPosition="left"
+        influencerWidth="narrow"
+        label={label}
+        onClose={onClose}
+        open={open}
+        size="wide"
+        title={title}
+        verticalPosition={verticalPosition}
+        ref={ref}>
+        <div
+          className={`${blockClass}__multi-step-panel-content`}
+          onBlur={handleBlur}
+          ref={contentRef}>
+          {renderChildren(children)}
         </div>
-      </ReactResizeDetector>
+      </TearsheetShell>
     );
   }
 );