+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.stories.js
@@ -47,13 +47,15 @@ export default {
 // Test values for props.
 
 const actionBarItems = [1, 2, 3, 4].map((item) => ({
-  keuy: `a-key-${item}`,
+  key: `a-key-${item}`,
   renderIcon: Lightning16,
   iconDescription: `Action ${item}`,
 }));
 
+const actionBarOverflowLabel = 'Show further action bar items';
+
 const manyActionBarItems = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10].map((item) => ({
-  keuy: `a-key-${item}`,
+  key: `a-key-${item}`,
   renderIcon: Lightning16,
   iconDescription: `Action ${item}`,
 }));
@@ -77,6 +79,9 @@ const manyBreadcrumbItems = (
     <BreadcrumbItem href="#">Breadcrumb 8</BreadcrumbItem>
   </>
 );
+const breadcrumbOverflowLabel =
+  'Open and close additional breadcrumb item list.';
+
 const className = 'client-class-1 client-class-2';
 const dummyPageContent = (
   <Grid className={`${storyClass}__dummy-content`} narrow={true}>
@@ -107,29 +112,35 @@ const dummyPageContent = (
 );
 const pageActions = [
   {
+    key: 'secondary',
     kind: 'secondary',
     label: 'Secondary button',
     onClick: () => {},
   },
   {
+    key: 'primary',
     kind: 'primary',
     label: 'Primary button',
     onClick: () => {},
   },
 ];
+const pageActionsOverflowLabel = 'Page actions...';
 
 const manyPageActions = [
   {
-    kind: 'secondary',
+    key: '1',
+    kind: 'danger',
     label: 'Secondary 1',
     onClick: () => {},
   },
   {
+    key: '2',
     kind: 'secondary',
     label: 'Secondary 2',
     onClick: () => {},
   },
   {
+    key: '3',
     kind: 'primary',
     label: 'Primary',
     onClick: () => {},
@@ -144,6 +155,7 @@ const statusIndicator = (
 const subtitle = 'Optional subtitle if necessary';
 const longSubtitle =
   'Optional subtitle if necessary, which is very long in this case, but will need to be handled somehow. It just keeps going on and on and on and on and on.';
+// cspell: disable
 const summaryDetails = (
   <div style={{ display: 'flex' }}>
     <p
@@ -166,6 +178,7 @@ const summaryDetails = (
     </p>
   </div>
 );
+// cspell: enable
 const tabBar = (
   <Tabs>
     <Tab label="Tab 1" />
@@ -237,10 +250,13 @@ const Template = (args) => {
 export const AllAttributesSet = Template.bind({});
 AllAttributesSet.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   actionBarItems,
+  actionBarOverflowLabel,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   subtitle,
   availableSpace: summaryDetails,
   navigation: tabBar,
@@ -255,27 +271,32 @@ TitleOnly.args = {
   title,
 };
 
-export const TitleAndPagections = Template.bind({});
-TitleAndPagections.args = {
+export const TitleAndPageActions = Template.bind({});
+TitleAndPageActions.args = {
   title,
   pageActions,
+  pageActionsOverflowLabel,
 };
 
 export const BreadcrumbItemsAndTitle = Template.bind({});
 BreadcrumbItemsAndTitle.args = {
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
 };
 
 export const BreadcrumbItemsTitleAndPageActions = Template.bind({});
 BreadcrumbItemsTitleAndPageActions.args = {
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   pageActions,
+  pageActionsOverflowLabel,
 };
 
 export const BreadcrumbItemsTitleAndStatus = Template.bind({});
 BreadcrumbItemsTitleAndStatus.args = {
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   availableSpace: statusIndicator,
@@ -284,6 +305,7 @@ BreadcrumbItemsTitleAndStatus.args = {
 export const BreadcrumbItemsTitleTabs = Template.bind({});
 BreadcrumbItemsTitleTabs.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   navigation: tabBar,
@@ -292,6 +314,7 @@ BreadcrumbItemsTitleTabs.args = {
 export const BreadcrumbItemsTitleIconTabs = Template.bind({});
 BreadcrumbItemsTitleIconTabs.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   titleIcon: Bee24,
@@ -301,24 +324,29 @@ BreadcrumbItemsTitleIconTabs.args = {
 export const BreadcrumbItemsTitlePageActionsTabs = Template.bind({});
 BreadcrumbItemsTitlePageActionsTabs.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   navigation: tabBar,
 };
 
 export const BreadcrumbItemsTitlePageActionsTags = Template.bind({});
 BreadcrumbItemsTitlePageActionsTags.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   tags,
 };
 
 export const BreadcrumbItemsTitleTabsTags = Template.bind({});
 BreadcrumbItemsTitleTabsTags.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   navigation: tabBar,
@@ -330,18 +358,23 @@ export const BreadcrumbItemsActionBarTitlePageActionsTabsTags = Template.bind(
 );
 BreadcrumbItemsActionBarTitlePageActionsTabsTags.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   actionBarItems,
+  actionBarOverflowLabel,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   navigation: tabBar,
 };
 
 export const BreadcrumbItemsActionBar = Template.bind({});
 BreadcrumbItemsActionBar.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   actionBarItems,
+  actionBarOverflowLabel,
   preCollapseTitleRow: true,
   title,
 };
@@ -349,9 +382,11 @@ BreadcrumbItemsActionBar.args = {
 export const BreadcrumbItemsTitlePageActionsSubtitle = Template.bind({});
 BreadcrumbItemsTitlePageActionsSubtitle.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   subtitle,
 };
 
@@ -360,9 +395,11 @@ export const BreadcrumbItemsTitlePageActionsSummarydetailsTabs = Template.bind(
 );
 BreadcrumbItemsTitlePageActionsSummarydetailsTabs.args = {
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   title,
   pageActions,
+  pageActionsOverflowLabel,
   availableSpace: summaryDetails,
   navigation: tabBar,
 };
@@ -371,11 +408,14 @@ export const AllAttributesSetKeepsBreadcrumbTabs = Template.bind({});
 AllAttributesSetKeepsBreadcrumbTabs.args = {
   preventBreadcrumbScroll: true,
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   actionBarItems,
+  actionBarOverflowLabel,
   title,
   titleIcon: Bee24,
   pageActions,
+  pageActionsOverflowLabel,
   subtitle,
   availableSpace: summaryDetails,
   navigation: tabBar,
@@ -386,11 +426,14 @@ export const AllAttributesSetPreCollapseTitle = Template.bind({});
 AllAttributesSetPreCollapseTitle.args = {
   preCollapseTitleRow: true,
   background: true,
+  breadcrumbOverflowLabel,
   breadcrumbItems,
   actionBarItems,
+  actionBarOverflowLabel,
   title,
   titleIcon: Bee24,
   pageActions,
+  pageActionsOverflowLabel,
   subtitle,
   availableSpace: summaryDetails,
   navigation: tabBar,
@@ -402,8 +445,10 @@ LongValuesManyItems.args = {
   background: true,
   breadcrumbItems: manyBreadcrumbItems,
   actionBarItems: manyActionBarItems,
+  actionBarOverflowLabel,
   title: longTitle,
   pageActions: manyPageActions,
+  pageActionsOverflowLabel,
   subtitle: longSubtitle,
   availableSpace: summaryDetails,
   navigation: longTabBar,
@@ -439,6 +484,7 @@ export const AllAttributesWithSwitches = TemplateWithSwitchedArgs.bind({});
 AllAttributesWithSwitches.args = {
   actionBarItems,
   actionBarItemsSwitchedArg: true,
+  actionBarOverflowLabel,
   availableSpace: summaryDetails,
   availableSpaceSwitchedArg: true,
   background: true,
@@ -449,6 +495,7 @@ AllAttributesWithSwitches.args = {
   navigationSwitchedArg: true,
   pageActions,
   pageActionsSwitchedArg: true,
+  pageActionsOverflowLabel,
   preCollapseTitleRow: false,
   subtitle,
   subtitleSwitchedArg: true,
@@ -495,6 +542,7 @@ export const PageHeaderWithCarbonHeader =
 PageHeaderWithCarbonHeader.args = {
   actionBarItems,
   actionBarItemsSwitchedArg: true,
+  actionBarOverflowLabel,
   availableSpace: summaryDetails,
   availableSpaceSwitchedArg: true,
   background: true,
@@ -505,6 +553,7 @@ PageHeaderWithCarbonHeader.args = {
   navigationSwitchedArg: true,
   pageActions,
   pageActionsSwitchedArg: true,
+  pageActionsOverflowLabel,
   preCollapseTitleRow: false,
   subtitle,
   subtitleSwitchedArg: true,